gui_requirements.txt
requests>=2.28.0
websocket-client>=1.4.0
numpy>=1.21.0
pandas>=1.5.0

PyQt5>=5.15.9
pyqtgraph>=0.13.3
psutil>=5.9.6

ccxt>=4.0.0

python-dateutil>=2.8.0
pytz>=2022.1

colorama>=0.4.6
tqdm>=4.64.0

python-telegram-bot>=20.0
slack-sdk>=3.20.0

backtrader>=1.9.78.123
matplotlib>=3.6.0

pytest>=7.0.0
black>=22.0.0